Remove MockTokenTransferModule
