+++ crates/ibc/src/core/handler.rs
@@ -257,7 +257,6 @@ mod tests {
     use crate::core::timestamp::Timestamp;
     use crate::mock::client_state::MockClientState;
     use crate::mock::consensus_state::MockConsensusState;
-    use crate::mock::context::applications::MockTokenTransferModule;
     use crate::mock::context::MockContext;
     use crate::mock::header::MockHeader;
     use crate::mock::router::MockRouter;
@@ -622,7 +621,7 @@ mod tests {
         for test in tests {
             let res = match test.msg.clone() {
                 TestMsg::Ics26(msg) => dispatch(&mut ctx, &mut router, msg).map(|_| ()),
-                TestMsg::Ics20(msg) => send_transfer(&mut ctx, &mut MockTokenTransferModule, msg)
+                TestMsg::Ics20(msg) => send_transfer(&mut ctx, &mut DummyTransferModule, msg)
                     .map_err(|e: TokenTransferError| ChannelError::AppModule {
                         description: e.to_string(),
                     })
