--- crates/ibc/src/mock/context/applications.rs
@@ -1,90 +1,0 @@
-//! Application context implementations for `MockContext`
-
-use subtle_encoding::bech32;
-
-use crate::applications::transfer::context::cosmos_adr028_escrow_address;
-use crate::applications::transfer::context::TokenTransferExecutionContext;
-use crate::applications::transfer::context::TokenTransferValidationContext;
-use crate::applications::transfer::error::TokenTransferError;
-use crate::applications::transfer::PrefixedCoin;
-use crate::core::ics24_host::identifier::{ChannelId, PortId};
-use crate::Signer;
-
-pub struct MockTokenTransferModule;
-
-impl TokenTransferValidationContext for MockTokenTransferModule {
-    type AccountId = Signer;
-
-    fn get_port(&self) -> Result<PortId, TokenTransferError> {
-        Ok(PortId::transfer())
-    }
-
-    fn get_escrow_account(
-        &self,
-        port_id: &PortId,
-        channel_id: &ChannelId,
-    ) -> Result<Self::AccountId, TokenTransferError> {
-        let addr = cosmos_adr028_escrow_address(port_id, channel_id);
-        Ok(bech32::encode("cosmos", addr).into())
-    }
-
-    fn can_send_coins(&self) -> Result<(), TokenTransferError> {
-        Ok(())
-    }
-
-    fn can_receive_coins(&self) -> Result<(), TokenTransferError> {
-        Ok(())
-    }
-
-    fn send_coins_validate(
-        &self,
-        _from_account: &Self::AccountId,
-        _to_account: &Self::AccountId,
-        _coin: &PrefixedCoin,
-    ) -> Result<(), TokenTransferError> {
-        Ok(())
-    }
-
-    fn mint_coins_validate(
-        &self,
-        _account: &Self::AccountId,
-        _coin: &PrefixedCoin,
-    ) -> Result<(), TokenTransferError> {
-        Ok(())
-    }
-
-    fn burn_coins_validate(
-        &self,
-        _account: &Self::AccountId,
-        _coin: &PrefixedCoin,
-    ) -> Result<(), TokenTransferError> {
-        Ok(())
-    }
-}
-
-impl TokenTransferExecutionContext for MockTokenTransferModule {
-    fn send_coins_execute(
-        &mut self,
-        _from_account: &Self::AccountId,
-        _to_account: &Self::AccountId,
-        _coin: &PrefixedCoin,
-    ) -> Result<(), TokenTransferError> {
-        Ok(())
-    }
-
-    fn mint_coins_execute(
-        &mut self,
-        _account: &Self::AccountId,
-        _coin: &PrefixedCoin,
-    ) -> Result<(), TokenTransferError> {
-        Ok(())
-    }
-
-    fn burn_coins_execute(
-        &mut self,
-        _account: &Self::AccountId,
-        _coin: &PrefixedCoin,
-    ) -> Result<(), TokenTransferError> {
-        Ok(())
-    }
-}
