+++ crates/ibc/src/test_utils.rs
@@ -1,5 +1,11 @@
+use subtle_encoding::bech32;
 use tendermint::{block, consensus, evidence, public_key::Algorithm};
 
+use crate::applications::transfer::context::{
+    cosmos_adr028_escrow_address, TokenTransferExecutionContext, TokenTransferValidationContext,
+};
+use crate::applications::transfer::error::TokenTransferError;
+use crate::applications::transfer::PrefixedCoin;
 use crate::core::ics04_channel::acknowledgement::Acknowledgement;
 use crate::core::ics04_channel::channel::{Counterparty, Order};
 use crate::core::ics04_channel::error::{ChannelError, PacketError};
@@ -156,5 +162,82 @@ impl Module for DummyTransferModule {
         _relayer: &Signer,
     ) -> (ModuleExtras, Result<(), PacketError>) {
         (ModuleExtras::empty(), Ok(()))
+    }
+}
+
+impl TokenTransferValidationContext for DummyTransferModule {
+    type AccountId = Signer;
+
+    fn get_port(&self) -> Result<PortId, TokenTransferError> {
+        Ok(PortId::transfer())
+    }
+
+    fn get_escrow_account(
+        &self,
+        port_id: &PortId,
+        channel_id: &ChannelId,
+    ) -> Result<Self::AccountId, TokenTransferError> {
+        let addr = cosmos_adr028_escrow_address(port_id, channel_id);
+        Ok(bech32::encode("cosmos", addr).into())
+    }
+
+    fn can_send_coins(&self) -> Result<(), TokenTransferError> {
+        Ok(())
+    }
+
+    fn can_receive_coins(&self) -> Result<(), TokenTransferError> {
+        Ok(())
+    }
+
+    fn send_coins_validate(
+        &self,
+        _from_account: &Self::AccountId,
+        _to_account: &Self::AccountId,
+        _coin: &PrefixedCoin,
+    ) -> Result<(), TokenTransferError> {
+        Ok(())
+    }
+
+    fn mint_coins_validate(
+        &self,
+        _account: &Self::AccountId,
+        _coin: &PrefixedCoin,
+    ) -> Result<(), TokenTransferError> {
+        Ok(())
+    }
+
+    fn burn_coins_validate(
+        &self,
+        _account: &Self::AccountId,
+        _coin: &PrefixedCoin,
+    ) -> Result<(), TokenTransferError> {
+        Ok(())
+    }
+}
+
+impl TokenTransferExecutionContext for DummyTransferModule {
+    fn send_coins_execute(
+        &mut self,
+        _from_account: &Self::AccountId,
+        _to_account: &Self::AccountId,
+        _coin: &PrefixedCoin,
+    ) -> Result<(), TokenTransferError> {
+        Ok(())
+    }
+
+    fn mint_coins_execute(
+        &mut self,
+        _account: &Self::AccountId,
+        _coin: &PrefixedCoin,
+    ) -> Result<(), TokenTransferError> {
+        Ok(())
+    }
+
+    fn burn_coins_execute(
+        &mut self,
+        _account: &Self::AccountId,
+        _coin: &PrefixedCoin,
+    ) -> Result<(), TokenTransferError> {
+        Ok(())
     }
 }
